code/requirements.txt
requests==2.31.0
numpy>=1.23.5,<2.0.0
python-dotenv>=0.21.0
beautifulsoup4
langchain
openai
python-dotenv
langchain_openai
pandas
fuzzywuzzy
python-Levenshtein
tqdm
